separate wte4j-admin into web-fragments and add wte4j-admin-war as deployable
We decided to split the wte4j-admin module into two web-fragments, wte4j-admin-auth and wte4j-admin, in order to encapsulate different functionalities and add modularity to the project. This way, we can easily incorporate those fragment into a bigger project.
