+++ pom.xml
@@ -1,3 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
 <!--
 
     Copyright (C) 2015 Born Informatik AG (www.born.ch)
@@ -14,9 +15,7 @@
     See the License for the specific language governing permissions and
     limitations under the License.
 
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+--><project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>org.wte4j</groupId>
 	<artifactId>wte4j-parent</artifactId>
@@ -56,6 +55,8 @@
 	<modules>
 		<module>wte4j-core</module>
 		<module>wte4j-admin</module>
+    <module>wte4j-admin-war</module>
+    <module>wte4j-admin-auth</module>
 	</modules>
 
 	<properties>
