+++ wte4j-admin-auth/.gitignore
@@ -1,0 +1,1 @@
+/target/
