+++ wte4j-admin-auth/pom.xml
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="UTF-8"?>
+<?xml version="1.0"?>
 <!--
 
     Copyright (C) 2015 Born Informatik AG (www.born.ch)
@@ -16,17 +16,16 @@
     limitations under the License.
 
 -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-
-	<!-- POM file generated with GWT webAppCreator -->
+<project
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
+	xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
 	<modelVersion>4.0.0</modelVersion>
 	<parent>
 		<groupId>org.wte4j</groupId>
 		<artifactId>wte4j-parent</artifactId>
 		<version>0.0.1-SNAPSHOT</version>
 	</parent>
-	<artifactId>wte4j-admin</artifactId>
+	<artifactId>wte4j-admin-auth</artifactId>
 	<packaging>jar</packaging>
 
 	<properties>
@@ -49,7 +48,7 @@
 				<executions>
 					<execution>
 						<goals>
-							<goal>resources</goal>
+             			 	<goal>resources</goal>
 							<goal>generateAsync</goal>
 							<goal>i18n</goal>
 							<goal>compile</goal>
@@ -59,10 +58,9 @@
 				</executions>
 				<configuration>
 					<modules>
-						<module>org.wte4j.ui.Wte4jAdmin</module>
+						<module>org.wte4j.ui.auth.Wte4jAuth</module>
 					</modules>
-					<runTarget>WteAdmin.html</runTarget>
-					<i18nMessagesBundle>org.wte4j.ui.client.Labels</i18nMessagesBundle>
+					<i18nMessagesBundle>org.wte4j.ui.auth.Labels</i18nMessagesBundle>
 					<persistentunitcache>false</persistentunitcache>
 					<webappDirectory>${project.build.outputDirectory}/META-INF/resources</webappDirectory>
 				</configuration>
@@ -81,7 +79,6 @@
 			</dependency>
 		</dependencies>
 	</dependencyManagement>
-
 	<dependencies>
 		<!-- GWT -->
 
@@ -103,46 +100,6 @@
 			<scope>provided</scope>
 		</dependency>
 
-		<!-- wte4j -->
-
-		<dependency>
-			<groupId> ${project.parent.groupId}</groupId>
-			<artifactId>wte4j-core</artifactId>
-			<version>${project.version}</version>
-		</dependency>
-
-		<!-- databasese -->
-		<dependency>
-			<groupId>commons-dbcp</groupId>
-			<artifactId>commons-dbcp</artifactId>
-			<version>1.3</version>
-		</dependency>
-
-		<dependency>
-			<groupId>org.hsqldb</groupId>
-			<artifactId>hsqldb</artifactId>			
-		</dependency>
-
-		<!-- JSON encoder -->
-
-		<dependency>
-			<groupId>com.fasterxml.jackson.core</groupId>
-			<artifactId>jackson-annotations</artifactId>
-			<version>2.5.1</version>
-		</dependency>
-
-		<dependency>
-			<groupId>com.fasterxml.jackson.core</groupId>
-			<artifactId>jackson-core</artifactId>
-			<version>2.5.1</version>
-		</dependency>
-
-		<dependency>
-			<groupId>com.fasterxml.jackson.core</groupId>
-			<artifactId>jackson-databind</artifactId>
-			<version>2.5.1</version>
-		</dependency>
-
 		<!-- Spring -->
 
 		<dependency>
@@ -158,53 +115,12 @@
 
 		<dependency>
 			<groupId>org.springframework</groupId>
-			<artifactId>spring-jdbc</artifactId>
-		</dependency>
-
-		<dependency>
-			<groupId>org.springframework</groupId>
-			<artifactId>spring-orm</artifactId>
-		</dependency>
-
-		<dependency>
-			<groupId>org.springframework</groupId>
-			<artifactId>spring-tx</artifactId>
-		</dependency>
-
-		<dependency>
-			<groupId>org.springframework</groupId>
-			<artifactId>spring-aop</artifactId>
-		</dependency>
-
-		<dependency>
-			<groupId>org.springframework</groupId>
 			<artifactId>spring-web</artifactId>
 		</dependency>
 
 		<dependency>
 			<groupId>org.springframework</groupId>
 			<artifactId>spring-webmvc</artifactId>
-		</dependency>
-
-		<!-- File Upload -->
-
-		<dependency>
-			<groupId>commons-fileupload</groupId>
-			<artifactId>commons-fileupload</artifactId>
-			<version>1.3.1</version>
-		</dependency>
-
-		<dependency>
-			<groupId>commons-io</groupId>
-			<artifactId>commons-io</artifactId>
-			<version>2.4</version>
-		</dependency>
-
-		<!-- JPA Provider -->
-		
-		<dependency>
-			<groupId>org.apache.openjpa</groupId>
-			<artifactId>openjpa</artifactId>			
 		</dependency>
 
 		<!-- logging -->
@@ -218,5 +134,4 @@
 			<artifactId>spring-test</artifactId>
 		</dependency>
 	</dependencies>
-
 </project>
