+++ wte4j-admin-auth/src/main/java/org/wte4j/ui/auth/config/AuthApplicationConfig.java
