+++ wte4j-admin-auth/src/main/resources/org/wte4j/ui/auth/Labels.properties
@@ -1,0 +1,18 @@
+applicationTitle=Word Template Engine
+
+templateDocumentName=Document
+templateLanguage=Language
+templateEditedAt=Edited at
+templateEditor=Editor
+templateLockingUser=Locking User
+
+downloadTemplate=Download Template
+lockTemplate=Lock Template
+unlockTemplate=Unlock Template
+deleteTemplate=Delete Template
+updateTemplate=Update Template
+
+deleteConfirmation=Are you sure you want to delete {0}?
+
+submit=Submit
+cancel=Cancel
