+++ wte4j-admin-auth/src/main/resources/org/wte4j/ui/auth/client/LoginForm.ui.xml
