+++ wte4j-admin-war/.gitignore
@@ -1,0 +1,1 @@
+/target/
