+++ wte4j-admin-war/src/main/resources/META-INF/persistence.xml
