--- wte4j-admin/src/main/java/org/wte4j/ui/auth/WteAuth.gwt.xml
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (C) 2015 Born Informatik AG (www.born.ch)
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-            http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-
--->
-<module rename-to='WteAuth'>
-	<!-- Inherit the core Web Toolkit stuff. -->
-	<inherits name='com.google.gwt.user.User' />
-	<inherits name="com.google.gwt.i18n.I18N" />
-	<inherits name="com.google.gwt.resources.Resources" />
-
-	<!-- We need the JUnit module in the main module, -->
-	<!-- otherwise eclipse complains (Google plugin bug?) -->
-	<!-- <inherits name='com.google.gwt.junit.JUnit' /> -->
-
-	<!-- Inherit the default GWT style sheet. You can change -->
-	<!-- the theme of your GWT application by uncommenting -->
-	<!-- any one of the following lines. -->
-	<inherits name='com.google.gwt.user.theme.standard.Standard' />
-	<!-- <inherits name='com.google.gwt.user.theme.chrome.Chrome'/> -->
-	<!-- <inherits name='com.google.gwt.user.theme.dark.Dark'/> -->
-
-	<!-- Other module inherits -->
-
-	<!-- Specify the app entry point class. -->
-	<entry-point class='org.wte4j.ui.auth.client.WteAuth' />
-
-	<!-- Specify the paths for translatable code -->
-	<source path='client' />
-
-</module>
