--- wte4j-admin/src/main/java/org/wte4j/ui/auth/client/WteAuth.java
@@ -1,39 +1,0 @@
-/**
- * Copyright (C) 2015 Born Informatik AG (www.born.ch)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.wte4j.ui.auth.client;
-
-import com.google.gwt.core.client.EntryPoint;
-import com.google.gwt.user.client.ui.RootPanel;
-
-/**
- * Entry point classes define <code>onModuleLoad()</code>.
- */
-public class WteAuth implements EntryPoint {
-
-	/**
-	 * This is the entry point method.
-	 */
-	public void onModuleLoad() {		
-		
-		LoginPage loginPage = new LoginPage();
-		
-		RootPanel root= RootPanel.get();
-		root.add(loginPage);
-
-		loginPage.show();
-		
-	}
-}
