+++ wte4j-admin/src/main/resources/META-INF/web-fragment.xml
@@ -1,0 +1,82 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Copyright (C) 2015 Born Informatik AG (www.born.ch) Licensed under the 
+	Apache License, Version 2.0 (the "License"); you may not use this file except 
+	in compliance with the License. You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0 
+	Unless required by applicable law or agreed to in writing, software distributed 
+	under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES 
+	OR CONDITIONS OF ANY KIND, either express or implied. See the License for 
+	the specific language governing permissions and limitations under the License. -->
+<web-fragment xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xmlns="http://java.sun.com/xml/ns/javaee" xmlns:web="http://java.sun.com/xml/ns/javaee"
+	xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-fragment_3_0.xsd"
+	id="Wte4jAdmin" version="3.0">
+	<context-param>
+		<param-name>contextClass</param-name>
+		<param-value>org.springframework.web.context.support.AnnotationConfigWebApplicationContext</param-value>
+	</context-param>
+	<context-param>
+		<param-name>contextConfigLocation</param-name>
+		<param-value>org.wte4j.ui.server.config.CoreApplicationConfig</param-value>
+	</context-param>
+	<listener>
+		<listener-class>org.springframework.web.context.ContextLoaderListener</listener-class>
+	</listener>
+	<listener>
+		<listener-class>org.springframework.web.context.request.RequestContextListener</listener-class>
+	</listener>
+	<filter>
+		<filter-name>SessionUserFilter</filter-name>
+		<filter-class>org.wte4j.ui.server.filters.SessionContextFilter</filter-class>
+	</filter>
+	<filter-mapping>
+		<filter-name>SessionUserFilter</filter-name>
+		<url-pattern>/*</url-pattern>
+		<dispatcher>REQUEST</dispatcher>
+	</filter-mapping>
+	<security-constraint>
+		<web-resource-collection>
+			<web-resource-name>Wte Protected</web-resource-name>
+			<url-pattern>/Wte4jAdmin/*</url-pattern>
+		</web-resource-collection>
+		<auth-constraint>
+			<role-name>wte4j-admin</role-name>
+		</auth-constraint>
+	</security-constraint>
+	<security-role>
+		<role-name>wte4j-admin</role-name>
+	</security-role>
+	<servlet>
+		<servlet-name>templateService</servlet-name>
+		<servlet-class>org.wte4j.ui.server.services.TemplateRepositoryGWTAdapter</servlet-class>
+	</servlet>
+	<servlet-mapping>
+		<servlet-name>templateService</servlet-name>
+		<url-pattern>/Wte4jAdmin/templateService/*</url-pattern>
+	</servlet-mapping>
+	<servlet>
+		<servlet-name>remoteLogging</servlet-name>
+		<servlet-class>com.google.gwt.logging.server.RemoteLoggingServiceImpl</servlet-class>
+	</servlet>
+	<servlet-mapping>
+		<servlet-name>remoteLogging</servlet-name>
+		<url-pattern>/Wte4jAdmin/remote_logging</url-pattern>
+	</servlet-mapping>
+	<servlet>
+		<servlet-name>wteAdminRestServlet</servlet-name>
+		<servlet-class>org.springframework.web.servlet.DispatcherServlet</servlet-class>
+		<init-param>
+			<param-name>contextClass</param-name>
+			<param-value>
+            	org.springframework.web.context.support.AnnotationConfigWebApplicationContext
+        	</param-value>
+		</init-param>
+		<init-param>
+			<param-name>contextConfigLocation</param-name>
+			<param-value>org.wte4j.ui.server.config.RestServiceConfig</param-value>
+		</init-param>
+	</servlet>
+	<servlet-mapping>
+		<servlet-name>wteAdminRestServlet</servlet-name>
+		<url-pattern>/Wte4jAdmin/rest/*</url-pattern>
+	</servlet-mapping>
+</web-fragment>
