--- wte4j-admin/src/main/resources/log4j.properties
@@ -1,14 +1,0 @@
-### log4j logging properties
-
-# configuration of stdout appender
-log4j.appender.Console=org.apache.log4j.ConsoleAppender
-log4j.appender.Console.layout=org.apache.log4j.PatternLayout
-log4j.appender.Console.layout.ConversionPattern=%d{dd.MM HH:mm:ss} %-5p %c %x - %m%n
-
-
-### Logging Levels ###
-# log4j Levels: ALL | TRACE | DEBUG | INFO | WARN | ERROR | FATAL | OFF
-
-log4j.rootLogger=INFO, Console
-org.wte4j=INFO
-org.springframework=INFO
