--- wte4j-admin/src/main/resources/log4j2.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Configuration status="info" strict="true" name="default"
-	packages="org.apache.logging.log4j">
-	
-	<Appenders>
-		<Console name="Console" target="SYSTEM_OUT">
-			<PatternLayout pattern="%d{ISO8601} [%t] %-5level: %logger{36}: %msg%n" />
-		</Console>
-	</Appenders>
-
-	<Loggers>
-
-		<Root level="warn">
-			<AppenderRef ref="Console"/>
-		</Root>
-		<Logger name="org.wte4j" level="info"/>
-		<Logger name="org.hibernate" level="info"/>
-		<Logger name="org.springframework" level="info"/>
-	</Loggers>
-
-</Configuration>
