--- wte4j-admin/src/main/webapp/WEB-INF/web.xml
@@ -1,126 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (C) 2015 Born Informatik AG (www.born.ch)
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-            http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-
--->
-<web-app xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://java.sun.com/xml/ns/javaee" xmlns:web="http://java.sun.com/xml/ns/javaee" xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd" id="WteAdmin" version="3.0">
-  <context-param>
-    <param-name>contextClass</param-name>
-    <param-value>org.springframework.web.context.support.AnnotationConfigWebApplicationContext</param-value>
-  </context-param>
-  <context-param>
-    <param-name>contextConfigLocation</param-name>
-    <param-value>org.wte4j.ui.server.config.CoreApplicationConfig</param-value>
-  </context-param>
-  <listener>
-    <listener-class>org.springframework.web.util.Log4jConfigListener</listener-class>
-  </listener>
-  <listener>
-    <listener-class>org.springframework.web.context.ContextLoaderListener</listener-class>
-  </listener>
-  <listener>
-    <listener-class>org.springframework.web.context.request.RequestContextListener</listener-class>
-  </listener>
-  <filter>
-    <filter-name>SessionUserFilter</filter-name>
-    <filter-class>org.wte4j.ui.server.filters.SessionContextFilter</filter-class>
-  </filter>
-  <filter-mapping>
-    <filter-name>SessionUserFilter</filter-name>
-    <url-pattern>/*</url-pattern>
-    <dispatcher>REQUEST</dispatcher>
-  </filter-mapping>
-  <security-constraint>
-    <web-resource-collection>
-      <web-resource-name>Wte Protected</web-resource-name>
-      <url-pattern>/*</url-pattern>
-    </web-resource-collection>
-    <auth-constraint>
-      <role-name>wte4j-admin</role-name>
-    </auth-constraint>
-  </security-constraint>
-  <security-constraint>
-    <web-resource-collection>
-      <web-resource-name>Wte Auth</web-resource-name>
-      <url-pattern>/WteAuth/*</url-pattern>
-    </web-resource-collection>
-  </security-constraint>
-  <security-role>
-    <role-name>wte4j-admin</role-name>
-  </security-role>
-  <login-config>
-    <auth-method>FORM</auth-method>
-    <form-login-config>
-      <form-login-page>/WteAuth.html</form-login-page>
-      <form-error-page>/WteAuth.html</form-error-page>
-    </form-login-config>
-  </login-config>
-  <servlet>
-    <servlet-name>templateService</servlet-name>
-    <servlet-class>org.wte4j.ui.server.services.TemplateRepositoryGWTAdapter</servlet-class>
-  </servlet>
-  <servlet-mapping>
-    <servlet-name>templateService</servlet-name>
-    <url-pattern>/WteAdmin/templateService/*</url-pattern>
-  </servlet-mapping>
-  <servlet>
-    <servlet-name>remoteLogging</servlet-name>
-    <servlet-class>com.google.gwt.logging.server.RemoteLoggingServiceImpl</servlet-class>
-  </servlet>
-  <servlet-mapping>
-    <servlet-name>remoteLogging</servlet-name>
-    <url-pattern>/WteAdmin/remote_logging</url-pattern>
-  </servlet-mapping>
-  <servlet>
-    <servlet-name>wteAdminRestServlet</servlet-name>
-    <servlet-class>org.springframework.web.servlet.DispatcherServlet</servlet-class>
-    <init-param>
-      <param-name>contextClass</param-name>
-      <param-value>
-            	org.springframework.web.context.support.AnnotationConfigWebApplicationContext
-        	</param-value>
-    </init-param>
-    <init-param>
-      <param-name>contextConfigLocation</param-name>
-      <param-value>org.wte4j.ui.server.config.RestServiceConfig</param-value>
-    </init-param>
-  </servlet>
-  <servlet-mapping>
-    <servlet-name>wteAdminRestServlet</servlet-name>
-    <url-pattern>/WteAdmin/rest/*</url-pattern>
-  </servlet-mapping>
-  <servlet>
-    <servlet-name>wte-auth</servlet-name>
-    <servlet-class>org.springframework.web.servlet.DispatcherServlet</servlet-class>
-    <init-param>
-      <param-name>contextClass</param-name>
-      <param-value>
-            	org.springframework.web.context.support.AnnotationConfigWebApplicationContext
-        	</param-value>
-    </init-param>
-    <init-param>
-      <param-name>contextConfigLocation</param-name>
-      <param-value>org.wte4j.ui.auth.config.AuthApplicationConfig</param-value>
-    </init-param>
-  </servlet>
-  <servlet-mapping>
-    <servlet-name>wte-auth</servlet-name>
-    <url-pattern>/WteAuth/rest/*</url-pattern>
-  </servlet-mapping>
-  <welcome-file-list>
-    <welcome-file>WteAdmin.html</welcome-file>
-  </welcome-file-list>
-</web-app>
