--- wte4j-admin/src/main/webapp/WteAdmin.html
@@ -1,60 +1,0 @@
-<!--
-
-    Copyright (C) 2015 Born Informatik AG (www.born.ch)
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-            http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-
--->
-<!doctype html>
-<!-- The DOCTYPE declaration above will set the    -->
-<!-- browser's rendering engine into               -->
-<!-- "Standards Mode". Replacing this declaration  -->
-<!-- with a "Quirks Mode" doctype may lead to some -->
-<!-- differences in layout.                        -->
-
-<html>
-<head>
-<meta http-equiv="content-type" content="text/html; charset=UTF-8">
-<!--                                           -->
-<!-- Any title is fine                         -->
-<!--                                           -->
-<title>WTE Admin GUI</title>
-
-<!--                                           -->
-<!-- This script loads your compiled module.   -->
-<!-- If you add any GWT meta tags, they must   -->
-<!-- be added before this line.                -->
-<!--                                           -->
-<script type="text/javascript" language="javascript"
-	src="WteAdmin/WteAdmin.nocache.js"></script>
-</head>
-
-<!--                                           -->
-<!-- The body can have arbitrary html, or      -->
-<!-- you can leave the body empty if you want  -->
-<!-- to create a completely dynamic UI.        -->
-<!--                                           -->
-<body>
-	<div id="wte4j-admin">
-
-		<!-- RECOMMENDED if your web app will not function without JavaScript enabled -->
-		<noscript>
-			<div
-				style="width: 22em; position: absolute; left: 50%; margin-left: -11em; color: red; background-color: white; border: 1px solid red; padding: 4px; font-family: sans-serif">
-				Your web browser must have JavaScript enabled in order for this
-				application to display correctly.</div>
-		</noscript>
-	</div>
-
-</body>
-</html>
