--- wte4j-admin/src/test/java/org/wte4j/ui/server/config/StandaloneJPAConfigTest.java
@@ -1,60 +1,0 @@
-/**
- * Copyright (C) 2015 Born Informatik AG (www.born.ch)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.wte4j.ui.server.config;
-
-import static org.junit.Assert.*;
-
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-import javax.sql.DataSource;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Import;
-import org.springframework.jdbc.datasource.embedded.EmbeddedDatabase;
-import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseBuilder;
-import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseType;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-
-@RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration(classes = { StandaloneJPAConfigTest.TestConfiguration.class })
-public class StandaloneJPAConfigTest {
-    
-    @PersistenceContext(unitName="wte4j-templates")
-    EntityManager em;
-    
-    @Test
-    public void test() {
-        assertNotNull(em);
-    }
-
-    @Import(StandaloneJPAConfig.class)
-    public static class TestConfiguration {
-        
-        @Bean
-        @Qualifier("wte4j")
-        public DataSource dataSource(){
-            
-            EmbeddedDatabaseBuilder builder = new EmbeddedDatabaseBuilder();
-            EmbeddedDatabase db = builder.setType(EmbeddedDatabaseType.HSQL).build();
-            return db;
-            
-        }
-    }
-}
