--- wte4j-admin/src/test/resources/org/wte4j/ui/WteAdminJUnit.gwt.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module>
-  <!-- Inherit our applications main module.                      -->
-  <inherits name='org.wte4j.ui.WteAdmin' />
-</module>
